fix 0 vs undefined issue
ds_grid_get returns 0 when the index is not found in GMS 1. In GMS 2 it
returns undefined. Add ds_grid_get_compat to get the old GMS 1
behaviour. Ideally all grid reads should be replaced by this but it
seems to be several thousands so we skip this for now.